refactor: changed packages for persistence
