+++ app/src/main/java/sap/ass01/layered/persistence/repository/UserRepository.java
@@ -1,8 +1,7 @@
 package sap.ass01.layered.persistence.repository;
 
-import sap.ass01.layered.persistence.DTO.UserDTO;
+import sap.ass01.layered.persistence.dto.UserDTO;
 
-import java.io.File;
 import java.util.List;
 import java.util.Optional;
