+++ app/src/main/java/sap/ass01/layered/persistence/repository/factory/EBikeRepositoryFactory.java
@@ -1,8 +1,8 @@
 package sap.ass01.layered.persistence.repository.factory;
 
-import sap.ass01.layered.persistence.MongoDB.MongoEBikeRepository;
-import sap.ass01.layered.persistence.inMemory.InMemoryEBikeRepository;
-import sap.ass01.layered.persistence.mysql.MySqlEBikeRepository;
+import sap.ass01.layered.persistence.ebike.MongoEBikeRepository;
+import sap.ass01.layered.persistence.ebike.DiskEBikeRepository;
+import sap.ass01.layered.persistence.ebike.MySqlEBikeRepository;
 import sap.ass01.layered.persistence.repository.DatabaseType;
 import sap.ass01.layered.persistence.repository.EBikeRepository;
 
@@ -10,7 +10,7 @@ public class EBikeRepositoryFactory {
     public static EBikeRepository createRepository(DatabaseType dbType) {
         switch (dbType) {
             case IN_MEMORY:
-                return new InMemoryEBikeRepository();
+                return new DiskEBikeRepository();
             case MONGODB:
                 return new MongoEBikeRepository();
             case MYSQL:
