+++ app/src/main/java/sap/ass01/layered/persistence/repository/factory/UserRepositoryFactory.java
@@ -1,8 +1,8 @@
 package sap.ass01.layered.persistence.repository.factory;
 
-import sap.ass01.layered.persistence.MongoDB.MongoUserRepository;
-import sap.ass01.layered.persistence.inMemory.InMemoryUserRepository;
-import sap.ass01.layered.persistence.mysql.MySQLUserRepository;
+import sap.ass01.layered.persistence.user.MongoUserRepository;
+import sap.ass01.layered.persistence.user.DiskUserRepository;
+import sap.ass01.layered.persistence.user.MySQLUserRepository;
 import sap.ass01.layered.persistence.repository.DatabaseType;
 import sap.ass01.layered.persistence.repository.UserRepository;
 
@@ -10,7 +10,7 @@ public class UserRepositoryFactory {
     public static UserRepository createRepository(DatabaseType dbType) {
         switch (dbType) {
             case IN_MEMORY:
-                return new InMemoryUserRepository();
+                return new DiskUserRepository();
             case MONGODB:
                 return new MongoUserRepository();
             case MYSQL:
