+++ app/src/main/java/sap/ass01/layered/persistence/ride/MongoRideRepository.java
@@ -1,4 +1,4 @@
-package sap.ass01.layered.persistence.MongoDB;
+package sap.ass01.layered.persistence.ride;
 
 import com.mongodb.MongoWriteException;
 import com.mongodb.client.MongoClient;
@@ -7,9 +7,9 @@ import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.IndexOptions;
 import org.bson.Document;
-import sap.ass01.layered.persistence.DTO.EBikeDTO;
-import sap.ass01.layered.persistence.DTO.RideDTO;
-import sap.ass01.layered.persistence.DTO.UserDTO;
+import sap.ass01.layered.persistence.dto.EBikeDTO;
+import sap.ass01.layered.persistence.dto.RideDTO;
+import sap.ass01.layered.persistence.dto.UserDTO;
 import sap.ass01.layered.persistence.repository.RideRepository;
 
 import java.util.ArrayList;
